requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile --extra=dev --output-file=requirements.txt --resolver=backtracking pyproject.toml
#
absl-py==1.3.0
    # via
    #   mujoco
    #   tensorboard
antlr4-python3-runtime==4.9.3
    # via
    #   hydra-core
    #   omegaconf
attrs==22.2.0
    # via deeprl (pyproject.toml)
black==22.12.0
    # via deeprl (pyproject.toml)
cachetools==5.2.0
    # via google-auth
certifi==2022.12.7
    # via requests
cfgv==3.3.1
    # via pre-commit
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via black
cloudpickle==2.2.0
    # via
    #   gymnasium
    #   scalene
commonmark==0.9.1
    # via rich
cytoolz==0.12.1
    # via deeprl (pyproject.toml)
debugpy==1.6.5
    # via deeprl (pyproject.toml)
distlib==0.3.6
    # via virtualenv
filelock==3.9.0
    # via virtualenv
glfw==2.5.5
    # via mujoco
google-auth==2.15.0
    # via
    #   google-auth-oauthlib
    #   tensorboard
google-auth-oauthlib==0.4.6
    # via tensorboard
grpcio==1.51.1
    # via tensorboard
gymnasium[classic-control,mujoco]==0.27.0
    # via
    #   deeprl (pyproject.toml)
    #   pettingzoo
    #   shimmy
gymnasium-notices==0.0.1
    # via gymnasium
hydra-core==1.3.1
    # via deeprl (pyproject.toml)
identify==2.5.12
    # via pre-commit
idna==3.4
    # via requests
imageio==2.23.0
    # via gymnasium
importlib-metadata==6.0.0
    # via
    #   gymnasium
    #   markdown
importlib-resources==5.10.2
    # via hydra-core
isort==5.11.4
    # via deeprl (pyproject.toml)
jax-jumpy==0.2.0
    # via gymnasium
jinja2==3.1.2
    # via scalene
markdown==3.4.1
    # via tensorboard
markupsafe==2.1.1
    # via
    #   jinja2
    #   werkzeug
mujoco==2.3.1.post1
    # via gymnasium
mypy==0.991
    # via deeprl (pyproject.toml)
mypy-extensions==0.4.3
    # via
    #   black
    #   mypy
nodeenv==1.7.0
    # via pre-commit
numpy==1.24.1
    # via
    #   gymnasium
    #   imageio
    #   jax-jumpy
    #   mujoco
    #   pandas
    #   pettingzoo
    #   shimmy
    #   tensorboard
nvidia-cublas-cu11==11.10.3.66
    # via
    #   nvidia-cudnn-cu11
    #   torch
nvidia-cuda-nvrtc-cu11==11.7.99
    # via torch
nvidia-cuda-runtime-cu11==11.7.99
    # via torch
nvidia-cudnn-cu11==8.5.0.96
    # via torch
oauthlib==3.2.2
    # via requests-oauthlib
omegaconf==2.3.0
    # via hydra-core
packaging==22.0
    # via hydra-core
pandas==1.5.2
    # via torch-tb-profiler
pathspec==0.10.3
    # via black
pettingzoo[mpe]==1.22.3
    # via deeprl (pyproject.toml)
pillow==9.4.0
    # via imageio
platformdirs==2.6.2
    # via
    #   black
    #   virtualenv
pre-commit==2.21.0
    # via deeprl (pyproject.toml)
protobuf==3.20.3
    # via tensorboard
py-spy==0.3.14
    # via deeprl (pyproject.toml)
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pydantic==1.10.4
    # via deeprl (pyproject.toml)
pygame==2.1.3.dev8
    # via
    #   gymnasium
    #   pettingzoo
pygments==2.14.0
    # via rich
pynvml==11.4.1
    # via scalene
pyopengl==3.1.6
    # via mujoco
python-dateutil==2.8.2
    # via pandas
pytz==2022.7
    # via pandas
pyyaml==6.0
    # via
    #   omegaconf
    #   pre-commit
requests==2.28.1
    # via
    #   requests-oauthlib
    #   tensorboard
requests-oauthlib==1.3.1
    # via google-auth-oauthlib
rich==13.0.0
    # via scalene
rsa==4.9
    # via google-auth
ruff==0.0.211
    # via deeprl (pyproject.toml)
scalene==1.5.18
    # via deeprl (pyproject.toml)
shimmy==0.2.0
    # via gymnasium
six==1.16.0
    # via
    #   google-auth
    #   python-dateutil
tensorboard==2.11.0
    # via torch-tb-profiler
tensorboard-data-server==0.6.1
    # via tensorboard
tensorboard-plugin-wit==1.8.1
    # via tensorboard
tomli==2.0.1
    # via
    #   black
    #   mypy
toolz==0.12.0
    # via
    #   cytoolz
    #   deeprl (pyproject.toml)
torch==1.13.1
    # via deeprl (pyproject.toml)
torch-tb-profiler==0.4.1
    # via deeprl (pyproject.toml)
typing-extensions==4.4.0
    # via
    #   black
    #   gymnasium
    #   mypy
    #   pydantic
    #   rich
    #   torch
urllib3==1.26.13
    # via requests
virtualenv==20.17.1
    # via pre-commit
werkzeug==2.2.2
    # via tensorboard
wheel==0.38.4
    # via
    #   nvidia-cublas-cu11
    #   nvidia-cuda-runtime-cu11
    #   scalene
    #   tensorboard
zipp==3.11.0
    # via
    #   importlib-metadata
    #   importlib-resources

# The following packages are considered to be unsafe in a requirements file:
# setuptools
